requirements.txt
absl-py==2.1.0
accelerate==0.34.2
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argcomplete==3.4.0
array_record==0.5.1
asciitree==0.3.3
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
awscli==1.33.22
beautifulsoup4==4.12.3
bitsandbytes==0.41.0
blosc==1.11.2
boto==2.49.0
boto3==1.34.132
botocore==1.34.140
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.6
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
datasets==2.20.0
decorator==5.1.1
decord==0.6.0
deepspeed==0.14.4
diffusers==0.11.1
dill==0.3.8
dlimp @ git+https://github.com/kvablack/dlimp@ad72ce3a9b414db2185bc0b38461d4101a65477a
dm-tree==0.1.8
dnspython==2.6.1
docker-pycreds==0.4.0
docutils==0.16
draccus==0.8.0
einops==0.6.1
einops-exts==0.0.4
email_validator==2.2.0
environs==11.0.0
etils==1.5.2
exceptiongroup==1.2.1
executing==2.0.1
Farama-Notifications==0.0.4
fastapi==0.111.0
fastapi-cli==0.0.4
fasteners==0.19
filelock==3.14.0
flash-attn==2.5.5
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
fvcore==0.1.5.post20221221
gast==0.5.4
gcs-oauth2-boto-plugin==3.2
gdown==5.2.0
Geometry==0.0.23
gitdb==4.0.11
GitPython==3.1.43
google-apitools==0.5.32
google-auth==2.17.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
google-reauth==0.1.1
grpcio==1.64.1
gsutil==5.30
gymnasium==0.29.1
h11==0.14.0
h5py==3.11.0
hjson==3.1.0
httpcore==1.0.5
httplib2==0.20.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
humanize==4.9.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==7.2.1
importlib_resources==6.4.0
iopath==0.1.10
ipdb==0.13.13
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
json-numpy==2.0.0
jsonlines==4.0.0
keras==2.15.0
kiwisolver==1.4.5
libclang==18.1.1
lit==18.1.7
mani-skill2==0.5.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
mergedeep==1.3.4
ml-dtypes==0.2.0
mmcv-full==1.7.2
model-index==0.1.11
monotonic==1.6
mpi4py @ file:///home/conda/feedstock_root/build_artifacts/mpi4py_1709160976447/work
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
multiprocessing-logging==0.3.4
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
numcodecs==0.12.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
opendatalab==0.0.10
OpenEXR==3.2.4
openmim==0.3.9
openxlab==0.1.1
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.1.0
orjson==3.10.5
oss2==2.17.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
peft==0.11.1
petrel-oss-sdk===v2.2.1-2-g1505ef3-master
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
portalocker==2.10.0
promise==2.3
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
-e git+https://github.com/facebookresearch/pytorch3d.git@89653419d0973396f3eff1a381ba09a07fffc2ed#egg=pytorch3d
pytz==2023.4
pyu2f==0.1.5
PyYAML==6.0.1
pyyaml-include==1.4.1
regex==2024.5.15
requests==2.28.2
requests-oauthlib==2.0.0
retry_decorator==1.1.1
rich==13.4.2
rsa==4.7.2
Rtree==1.2.0
s3transfer==0.10.2
safetensors==0.4.3
sapien==2.2.2
scikit-learn==1.5.0
scipy==1.13.1
sentencepiece==0.1.99
sentry-sdk==2.6.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-graphics==2021.12.3
tensorflow-hub==0.16.1
tensorflow-io==0.37.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-metadata==1.15.0
tensorflow-probability==0.22.0
termcolor==2.4.0
threadpoolctl==3.5.0
timm==0.9.10
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.1
torch==2.2.0
torchprofile==0.0.4
torchvision==0.17.0
tqdm==4.65.2
traitlets==5.14.3
transformers==4.46.3
transforms3d==0.4.2
trimesh==4.4.1
triton==2.2.0
typeguard==2.13.3
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.19
uvicorn==0.30.1
uvloop==0.19.0
wandb==0.18.5
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wrapt==1.14.1
xformers==0.0.24
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zarr==2.18.2
zipp==3.19.2
